添加日志配置文件并在代码中加入日志

+++ main.py
@@ -2,24 +2,22 @@ import os
 import base64
 from pathlib import Path
 from configparser import ConfigParser
-import logging
-from logging.handlers import RotatingFileHandler
+import logging.config
 
 from wallpaper_downloader import WallpaperDownloader
-from utils.database import WallpaperDatabase, ImageInfo, FULLSTARTDATE, ENDDATE, URL, COPYRIGHT, COPYRIGHTLINK, TITLE
+from utils.database import WallpaperDatabase
 from utils.email import send_email
 
-rotating_file_handler = RotatingFileHandler(f"{__file__[:-3]}.log",
-                                            maxBytes=1024 * 1024 * 4, backupCount=3, encoding='utf-8')
-rotating_file_handler.setFormatter(logging.Formatter('%(asctime)s[%(levelname)s]%(filename)s(%(lineno)d): %(message)s'))
-
-logger = logging.getLogger(__name__)
-logger.setLevel(logging.DEBUG)
-logger.addHandler(rotating_file_handler)
-
+SCRIPT_NAME = 'DLBINGWALLPAPER'
 CONFIG_FILE = 'config.ini'
+LOGGING_CONF = 'logging.conf'
+
+logging.config.fileConfig(f'./{LOGGING_CONF}')
+logger = logging.getLogger('main')
+
 
 def main():
+    logger.info(f"{SCRIPT_NAME} start running!")
     downloader = WallpaperDownloader()
     database = WallpaperDatabase(os.path.abspath(os.curdir))
     conf = ConfigParser()
@@ -52,22 +50,22 @@ def main():
         return
 
     info_list: list = downloader.image_info_list(day_count=8)
-    logger.debug(info_list)
-
-    SMTP_PORT = 25 #587 
 
     with database.open_db_context():
         # 下载 & 保存
-        
         for i, info in enumerate(info_list):
-            logger.info(f"downloading: {info.copyright}")
+            logger.info(f"downloading: {info.enddate} {info.copyright}")
+            if database.record_exist(info.enddate):
+                logger.info('wallpaper exist, skip.')
+                continue
+
             image_file = Path(save_path) / f"{info.enddate}_{info.title}.jpg"
             downloader.download_image(info.url, image_file)
             data = image_file.read_bytes()
             b64_data = base64.b64encode(data)
             if database.save_info(info, b64_data):
                 send_email(smtp_host, smtp_port, from_mailbox, auth_password, to_mailboxes, f"[Bing今日美图] {info.title}", (image_file, ))
-                pass
+                logger.info("email has been sent to the specified mailbox.")
             
             if not save_as_image_file:
                 image_file.unlink(missing_ok=True)
@@ -90,5 +88,6 @@ def main():
         image_file = Path(os.curdir) / "20240621.jpg"
         image_file.write_bytes(image_data)
         '''
+    logger.info(f"{SCRIPT_NAME} exit.")
 
 main()
